requirements.txt
py-algorand-sdk>=2.6.0
pyteal>=0.25.0
algokit-utils>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pycryptodome>=3.20.0
twilio>=8.13.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
